API

ReefModEngine.DataCube — Method

DataCube(data::AbstractArray; kwargs...)::YAXArray

Constructor for YAXArray. When used with axes_names, the axes labels will be UnitRanges from 1 up to that axis length.

Arguments

data : Array of data to be used when building the YAXArray

axes_names :

ReefModEngine.area_needed — Method

area_needed(n_corals::Int64, density::Float64)::Float64

Determine area needed to deploy the given number of corals at the specified density.

ReefModEngine.deployment_area — Method

deployment_area(n_corals::Int64, max_n_corals::Int64, density::Float64, target_areas::Vector{Float64})::Tuple{Float64,Float64}

Determine deployment area for the expected number of corals to be deployed.

Arguments

n_corals : Number of corals,

max_n_corals : Expected maximum deployment effort (total number of corals in intervention set)

density : Stocking density

target_areas : Available area at target location(s)

Returns

Tuple

Percent area of deployment

modified stocking density

ReefModEngine.init_rme — Method

init_rme(rme_path::String)::Nothing

Initialize ReefMod Engine for use.

ReefModEngine.match_id — Method

match_id(id::String)::Int64
match_ids(ids::Vector{String})::Vector{Int64}

Find matching index position for the given ID(s) according to ReefMod Engine's reef list.

Note

ReefMod Engine's reef list is in all upper case. The provided IDs are converted to upper case to ensure a match.

Examples

julia> reef_ids()
# 3806-element Vector{String}:
#  "10-330"
#  "10-331"
#  ⋮
#  "23-048"
#  "23-049"

julia> match_id("10-330")
#  1

julia> match_id("23-049")
#  3806

julia> match_ids(["23-048", "10-331"])
#  3805
#  2

ReefModEngine.reef_ids — Method

reef_ids()::Vector{String}

Get list of reef ids in the order expected by ReefMod Engine.

ReefModEngine.reset_rme — Method

reset_rme()

Reset ReefModEngine, clearing any and all interventions and reef sets.

ReefModEngine.set_enrichment_deployment! — Method

set_enrichment_deployment!(
    name::String,
    reefset::String,
    n_larvae::Int64,
    year::Int64,
    area_km2::Vector{Float64},
    density::Float64
)::Nothing

As set_seeding_deployment() but for larvae enrichment (also known as assisted migration). Set deployment for a single target year.

set_enrichment_deployment!(
    name::String,
    reefset::String,
    n_larvae::Int64,
    max_effort::Int64,
    first_year::Int64,
    last_year::Int64,
    year_step::Int64,
    area_km2::Vector{Float64},
    density::Float64
)::Nothing

Set deployment for multiple years at a given frequency.

ReefModEngine.set_outplant_deployment! — Method

set_outplant_deployment!(name::String, reefset::String, n_corals::Int64, year::Int64, area_km2::Vector{Float64}, density::Float64)::Nothing

Set outplanting deployments for a single year.

Arguments

name : Name to assign intervention event

reefset : Name of pre-defined list of reefs to intervene on

n_corals : Number of corals to outplant

year : Year to intervene

area_km2 : Area to intervene [km²]

density : Stocking density of intervention [corals / m²]

setoutplantdeployment!(name::String, reefset::String, ncorals::Int64, maxeffort::Int64, firstyear::Int64, lastyear::Int64, yearstep::Int64, areakm2::Vector{Float64}, density::Float64)::Nothing

Set outplanting deployments across a range of years.

Arguments

name : Name to assign intervention event

reefset : Name of pre-defined list of reefs to intervene on

n_corals : Number of corals to outplant for a given year

max_effort : Total number of corals to outplant

first_year : First year to start interventions

last_year : Final year of interventions

year_step : Frequency of intervention (1 = every year, 2 = every second year, etc)

area_km2 : Intervention area [km²]

density : Stocking density of intervention [corals / m²]